requirements.txt
fastmcp>=0.2.0
python-frontmatter>=1.0.0
GitPython>=3.1.40
